On Windows, implements JNI-based replacements for the JNA-based H.264 encoder and decoder. The new implementations execute much faster, produce less garbage and use a newer ffmpeg version (which contains optimizations to at least the H.264 decoder).
Mac OS X and Linux will still use the JNA-based implementations because the respective native JNI libraries haven't been built for them.

Just as the JNA-based implementations, the JNI-based ones cause disturbing numbers of garbage collections (though less than their ancestors) - the encoder and the packetizer are huge offenders. But these will be addressed in subsequent revisions because the problems are not the subject of the switch from JNA to JNI.

git-svn-id: https://svn.java.net/svn/jitsi~svn/trunk@4922 198c5059-ef06-4ec0-9a75-fdd6b0ef00a6