Add ZID name feature.
The ZID name feature enables the user to associate a name for a ZRTP
identifier. The user can enter such a name only if the authentication string
(SAS) was compared and confirmed.

The popup dialog shows the current ZID name or is preset with the display
name of the called party in case no name was associated previously. The user
may change the name and store it with OK, pressing Cancel leaves the name
untouched.

git-svn-id: https://svn.java.net/svn/jitsi~svn/trunk@9515 198c5059-ef06-4ec0-9a75-fdd6b0ef00a6