Corrects parsing of Swift XMPP entity capabilities response.
Problem:
Swift sends entity capabilities responses (XEP-0115 Disco#info responses)
without adding a "node" attribute inside the "query" element. This is a correct
behavior as the "node" attribute is optional (XEP-0030). But Jitsi was checking
the presence of this attribute and was rejecting all DiscoveryInfo without it,
but containing a defined "hash" (for filtering Google Talk web, which does not
send "hash").

Correction:
Removes the check of the "node" attribute to validate Disco#info responses.

Bug report:
Kim Alvefur: [jitsi-dev] Invalid DiscoverInfo


git-svn-id: https://svn.java.net/svn/jitsi~svn/trunk@9588 198c5059-ef06-4ec0-9a75-fdd6b0ef00a6