Adds an "answer/hangup" call global key shortcut. Modifies the "hangup" global key shortcut behavior.
The answer/hangup shortcut:
--------------------------
This shortcut may be helpful for headsets with a single button for answering or
hanging up a call.

By default this shortcut is set to "Ctrl + Shift + p". When pressed:
1) If there is at least an incoming call, then Jitsi answers the first incoming
call.
2) If there is no incoming call, then Jitsi closes all active calls (An active
    call is a call with at least one peer which is not on hold).
3) If there is no incoming call and no active call, then Jitsi closes all
remaining calls (which are all on hold).


New hangup shortcut behavior:
-----------------------------
The hangup global key shortcut behavior is modified as follow:
1) If there is at least an incoming call, then Jitsi closes the first incoming
call.
2) If there is no incoming call, then Jitsi closes all active calls (An active
    call is a call with at least one peer which is not on hold).
3) If there is no incoming call and no active call, then Jitsi closes all
remaining calls (which are all on hold).


git-svn-id: https://svn.java.net/svn/jitsi~svn/trunk@9591 198c5059-ef06-4ec0-9a75-fdd6b0ef00a6